Change accrual type

+++ internal/core/domain/order.go
@@ -16,7 +16,7 @@ type Order struct {
 	UserID      int         `db:"user_id" json:"-"`
 	OrderNumber int         `db:"order_number" json:"number"`
 	Status      OrderStatus `db:"status" json:"status"`
-	Accrual     int         `db:"accrual" json:"accrual,omitempty"`
+	Accrual     float64     `db:"accrual" json:"accrual,omitempty"`
 	CreatedAt   time.Time   `db:"created_at" json:"uploaded_at"`
 	UpdatedAt   time.Time   `db:"updated_at" json:"-"`
 }
