+++ internal/core/services/accrualservice/accrualservice.go
@@ -51,9 +51,9 @@ func (a *AccrualService) Stop() {
 }
 
 type AccrualResponse struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
 
 func (a *AccrualService) checkOrders() {
