+++ migrations/000003_add_accrual_to_order.up.sql
@@ -1,2 +1,2 @@
 alter table orders
-add column accrual int;
+add column accrual numeric;
